add diagram for relationship tables, move to DataRepoExpressionFixture
